python/requirements.txt
openai>=2.6.1
python-dotenv>=1.0.0
pydantic>=2.0.0
